w105 ztag

+++ legacy_one/trunk/src/f_finale.c
@@ -330,7 +330,7 @@ void F_Ticker (void)
 
             if( gamemode != doom2_commercial)
             {
-                ULONG f = finalecount;
+                uint32_t  f = finalecount;
                 if( f >= MAXINT/2 )
                     f -= MAXINT/2;
